strarray: move to its own file
